ci: replace nv-base skill gate with playbook compliance
